feat: 상담신청서 REST API 개발완료
Former-commit-id: 60de25de9907bd324f837a3db9de3b291e8e21b7

+++ backend/src/main/java/com/B204/lawvatar_backend/application/dto/GetMyApplicationListResponse.java
@@ -22,6 +22,6 @@ public class GetMyApplicationListResponse {
     String recommendedQuestions;
     boolean isCompleted;
     LocalDateTime createdAt;
-    List<Integer> tags;
+    List<Long> tags;
 
 }
